backend/requirements.txt
Django==3.1.2
psycopg2==2.8.5
djangorestframework==3.12.0
django-cors-headers==3.2.1
gunicorn==20.0.4
